move pre pull to settings as well
