+++ services/autoscaling/src/simcore_service_autoscaling/utils/auto_scaling_core.py
@@ -8,7 +8,7 @@ from models_library.generated_models.docker_rest_api import Node
 from types_aiobotocore_ec2.literals import InstanceTypeType
 
 from ..core.errors import Ec2InstanceInvalidError, Ec2InvalidDnsNameError
-from ..core.settings import ApplicationSettings
+from ..core.settings import EC2InstanceBootSpecific
 from ..models import (
     AssignedTasksToInstance,
     AssignedTasksToInstanceType,
@@ -74,11 +74,9 @@ async def associate_ec2_instances_with_nodes(
     return associated_instances, non_associated_instances
 
 
-async def ec2_startup_script(app_settings: ApplicationSettings) -> str:
+async def ec2_startup_script(ec2_boot_specific: EC2InstanceBootSpecific) -> str:
     assert app_settings.AUTOSCALING_EC2_INSTANCES  # nosec
-    startup_commands = (
-        app_settings.AUTOSCALING_EC2_INSTANCES.EC2_INSTANCES_CUSTOM_BOOT_SCRIPTS.copy()
-    )
+    startup_commands = ec2_boot_specific.custom_boot_scripts.copy()
     startup_commands.append(await utils_docker.get_docker_swarm_join_bash_command())
     if app_settings.AUTOSCALING_REGISTRY:  # noqa: SIM102
         if pull_image_cmd := utils_docker.get_docker_pull_images_on_start_bash_command(
